+++ msc-stealth-login.php
@@ -1,9 +1,9 @@
 <?php
 /**
  * Plugin Name: MSC Stealth Login
- * Plugin URI: https://anomalous.co.za
+ * Plugin URI: https://github.com/djm56/msc-stealth-login
  * Description: Hide your login page, block brute force attacks, and protect your WordPress site from unauthorized access. Complete free plugin with all features included.
- * Version: 1.0.2
+	 * Version: 1.0.4
  * Author: Anomalous Developers
  * Author URI: https://anomalous.co.za
  * Text Domain: msc-stealth-login
@@ -22,7 +22,7 @@ if ( ! defined( 'ABSPATH' ) ) {
 /**
  * Current plugin version.
  */
-define( 'MSCSL_PLUGIN_VERSION', '1.0.2' );
+define( 'MSCSL_PLUGIN_VERSION', '1.0.4' );
 
 /**
  * Absolute path to the main plugin file.